moved dynamic ticker into types section
